ClusterService discovered here. Module `domain`.
Module `domain` is the most important part with definitions of domain
models. This is my first DDD tries. ClusterService is domain service
for do correlations. Now this is so much simplified, that is why i.e.
method `match_id` from `domain.Match` returns `match_portal_id`.

Here we start use attrs. And this will be continued.